Reorganized matches.py
